New drivers with lots of comments and explanation on how to contribute to SVP. (new Opal-RT drivers and PHIL model)

+++ Lib/svpelab/dataset.py
@@ -29,7 +29,9 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 Questions can be directed to support@sunspec.org
 """
-
+import datetime
+import os
+import pandas as pd
 
 class DatasetError(Exception):
     """
@@ -58,33 +60,56 @@ class DatasetError(Exception):
     A dataset consists of a set of time series points organized as parallel arrays and some
     additional optional properties.
 
-    Optional roperties:
+    Optional properties:
     Start time of dataset
     Sample rate of dataset (samples/sec)
     Trigger sample (record index into dataset)
 
 """
 class Dataset(object):
-    def __init__(self, points=None, data=None, start_time=None, sample_rate=None, trigger_sample=None, params=None):
+    def __init__(self, points=None, data=None, start_time=None, sample_rate=None, trigger_sample=None, params=None,
+                 ts=None):
         self.start_time = start_time              # start time
         self.sample_rate = sample_rate            # samples/second
         self.trigger_sample = trigger_sample      # trigger sample
         self.points = points                      # point names
         self.data = data                          # data
+        self.ts = ts
+        self.df = None
 
         if points is None:
             self.points = []
         if data is None:
             self.clear()
 
+    def point_data(self, point):
+        try:
+            idx = self.points.index(point)
+        except ValueError:
+            raise DatasetError('Data point not in dataset: %s' % point)
+        return self.data[idx]
+
     def append(self, data):
         dlen = len(data)
+        # self.ts.log_debug('self.data=%s, data=%s' % (self.data, data))
         if len(data) != len(self.data):
             raise DatasetError('Append record point mismatch, dataset contains %s points,'
                                ' appended data contains %s points' % (len(self.data), dlen))
         for i in range(dlen):
             try:
-                v = float(data[i])
+                if data[i] is not None:
+                    if data[i] is tuple:
+                        self.ts.log_debug('tuple data point recorded: %s' % data)
+                        v = float(data[i][0])
+                    elif isinstance(data[i], datetime.datetime):
+                        epoch = datetime.datetime.utcfromtimestamp(0)
+                        total_seconds = (data[i] - epoch).total_seconds()
+                        # total_seconds will be in decimals (millisecond precision)
+                        v = total_seconds
+                    else:
+                        v = float(data[i])
+                else:
+                    v = 'None'
             except ValueError:
                 v = data[i]
             self.data[i].append(v)
@@ -103,33 +128,59 @@ class Dataset(object):
             self.data.append([])
 
     def to_csv(self, filename):
-        cols = range(len(self.data))
-        if len(cols) > 0:
-            f = open(filename, 'w')
-            f.write('%s\n' % ', '.join(map(str, self.points)))
-            for i in xrange(len(self.data[0])):
-                d = []
-                for j in cols:
-                    d.append(self.data[j][i])
-                f.write('%s\n' % ', '.join(map(str, d)))
-            f.close()
+        """
+        Write result csv file based on the dataset. If the Simulation csv mode is used for the DAS, the csv is written
+        through a pandas dataset.
+
+        :param filename: String Path and name of the csv file to write
+
+        :return: nothing
+        """
+        mode = self.ts.param_value('das.' + 'mode')
+        if mode != 'DAS Simulation' or self.df is None:
+            cols = list(range(len(self.data)))
+            if len(cols) > 0:
+                f = open(filename, 'w')
+                f.write('%s\n' % ', '.join(map(str, self.points)))
+                for i in range(len(self.data[0])):
+                    d = []
+                    for j in cols:
+                        # self.ts.log_debug('data = %s' % self.data)
+                        # self.ts.log_debug('point names = %s' % self.points)
+                        # self.ts.log_debug('len(points) = %s, len(data) = %s' % (len(self.points), len(self.data)))
+                        # self.ts.log_debug('j = %s, i = %i, self.data[j][i] = %s' % (j, i, self.data[j][i]))
+                        d.append(self.data[j][i])
+                    f.write('%s\n' % ', '.join(map(str, d)))
+                f.close()
+        else:
+            self.df.to_csv(filename, index=False)
 
     def from_csv(self, filename, sep=','):
         self.clear()
         f = open(filename, 'r')
-        ids = None
-        while ids is None:
-            line = f.readline().strip()
-            if len(line) > 0 and line[0] != '#':
-                ids = [e.strip() for e in line.split(sep)]
-        self.points = ids
-        for i in range(len(self.points)):
-            self.data.append([])
-        for line in f:
-            data = [float(e.strip()) for e in line.split(sep)]
-            if len(data) > 0:
-                self.append(data)
+        lines = f.readlines()
+        self.points = [e.strip() for e in lines.pop(0).split(sep)]
+        for line in lines:
+            data = []
+            for e in line.split(sep):
+                if 'Step' not in e and 'None' not in e:
+                    data.append(float(e.strip()))
+                elif 'Step' in e or 'None' in e:
+                    data.append(e.strip())
+                    if len(data) > 0:
+                        self.data.append(data)
+                # else:
+                #     data.append(e.strip())
+
         f.close()
+    def remove_none_row(self,filename, index):
+        import pandas as pd
+        import numpy as np
+        df = pd.read_csv(filename)
+        df[index].replace('None', np.nan, inplace=True)
+        df.dropna(subset=[index],inplace=True)
+        df.reset_index(inplace=True,drop=True)
+        df.to_csv(filename,index=False)
 
 
 if __name__ == "__main__":
